TASK-29 toolbar-editing.component.tsx implemented
